Auto lookup failure handler follows _AUTO setting
It appears that the failure handler in i18n plugin (l. 175) defaults to
hard-coded behaviour without possibility to change the "fail" attribute
or use the _AUTO feature¹ (which has no effect actually).

    $handle->fail_with(sub { $_[1] });

This fixup restores the Locale::Maketext handler which dies if no entry
key is found in lexicon. Legacy code is not broken if default _AUTO
snippet was kept.

Also, I propose to display missing entries in debug mode.

¹ http://mojolicious.org/perldoc/Locale::Maketext#CONTROLLING-LOOKUP-FAILURE